requirements/base.txt
# Base
pytz==2023.3
python-slugify==6.1.2
psycopg2==2.9.3 --no-binary psycopg2

# Django
Django==4.0
django-cors-headers==3.10.1
django-polymorphic==3.1.0
djangorestframework-simplejwt==5.2.2
djangorestframework-simplejwt[crypto]

# Environment
django-environ==0.9.0

# Pictures
Pillow

# Security
argon2-cffi==21.3.0
PyJWT==v1.7.1

# Static files
whitenoise==5.3.0

# Extras
unidecode==1.3.6

# Celery
django-redis==5.2.0
celery==5.1.2
flower==1.2.0
redis==4.3.4

# restframework
djangorestframework-camel-case==1.3.0
djangorestframework==3.14.0
drf-schema-adapter==3.0.6
django-filter==23.2

# Documentation
drf-yasg==1.21.7

